refactoring: decouple clickhouse into browse.ts and node.ts
