+++ packages/common-utils/src/clickhouse/index.ts
@@ -10,15 +10,14 @@ import { isSuccessfulResponse } from '@clickhouse/client-common';
 import * as SQLParser from 'node-sql-parser';
 import objectHash from 'object-hash';
 
+import { Metadata } from '@/metadata';
 import {
   renderChartConfig,
   setChartSelectsAlias,
   splitChartConfigs,
 } from '@/renderChartConfig';
 import { ChartConfigWithOptDateRange, SQLInterval } from '@/types';
-import { hashCode, isBrowser, isNode, timeBucketByGranularity } from '@/utils';
-
-import { Metadata } from './metadata';
+import { hashCode } from '@/utils';
 
 // export @clickhouse/client-common types
 export type {
@@ -357,34 +356,7 @@ export const computeResultSetRatio = (resultSet: ResponseJSON<any>) => {
   return result;
 };
 
-const localModeFetch: typeof fetch = (input, init) => {
-  if (!init) init = {};
-  const url = new URL(
-    input instanceof URL ? input : input instanceof Request ? input.url : input,
-  );
-
-  // CORS is unhappy with the authorization header, so we will supply as query params instead
-  const auth: string = init.headers?.['Authorization'];
-  const [username, password] = window
-    .atob(auth.substring('Bearer'.length))
-    .split(':');
-  delete init.headers?.['Authorization'];
-  delete init.headers?.['authorization'];
-  if (username) url.searchParams.set('user', username);
-  if (password) url.searchParams.set('password', password);
-
-  return fetch(`${url.toString()}`, init);
-};
-
-const standardModeFetch: typeof fetch = (input, init) => {
-  if (!init) init = {};
-  // authorization is handled on the backend, don't send this header
-  delete init.headers?.['Authorization'];
-  delete init.headers?.['authorization'];
-  return fetch(input, init);
-};
-
-interface QueryInputs<Format extends DataFormat> {
+export interface QueryInputs<Format extends DataFormat> {
   query: string;
   format?: Format;
   abort_signal?: AbortSignal;
@@ -398,39 +370,31 @@ export type ClickhouseClientOptions = {
   host: string;
   username?: string;
   password?: string;
+  getAccessToken?: () => Promise<string>;
 };
 
-export const getJSNativeCreateClient = async () => {
-  if (isBrowser) {
-    // Only import client-web in browser environment
-    const { createClient } = await import('@clickhouse/client-web');
-    return createClient;
-  } else if (isNode) {
-    // Use require with eval to prevent webpack from analyzing this import
-    // This ensures @clickhouse/client is never bundled in browser builds
-    const { createClient } = eval('require')(
-      '@clickhouse/client',
-    ) as typeof import('@clickhouse/client');
-    return createClient;
-  }
-  throw new Error('Unsupported environment');
-};
-
-export class ClickhouseClient {
-  private readonly host: string;
-  private readonly username?: string;
-  private readonly password?: string;
+export abstract class BaseClickhouseClient {
+  protected readonly host: string;
+  protected readonly username?: string;
+  protected readonly password?: string;
+  protected readonly getAccessToken?: () => Promise<string>;
   /*
    * Some clickhouse db's (the demo instance for example) make the
    * max_rows_to_read setting readonly and the query will fail if you try to
    * query with max_rows_to_read specified
    */
-  private maxRowReadOnly: boolean;
+  protected maxRowReadOnly: boolean;
 
-  constructor({ host, username, password }: ClickhouseClientOptions) {
+  constructor({
+    host,
+    username,
+    password,
+    getAccessToken,
+  }: ClickhouseClientOptions) {
     this.host = host;
     this.username = username;
     this.password = password;
+    this.getAccessToken = getAccessToken;
     this.maxRowReadOnly = false;
   }
 
@@ -479,111 +443,9 @@ export class ClickhouseClient {
     throw new Error('ClickHouseClient query impossible codepath');
   }
 
-  // https://github.com/ClickHouse/clickhouse-js/blob/1ebdd39203730bb99fad4c88eac35d9a5e96b34a/packages/client-web/src/connection/web_connection.ts#L151
-  private async __query<Format extends DataFormat>({
-    query,
-    format = 'JSON' as Format,
-    query_params = {},
-    abort_signal,
-    clickhouse_settings: external_clickhouse_settings,
-    connectionId,
-    queryId,
-  }: QueryInputs<Format>): Promise<BaseResultSet<ReadableStream, Format>> {
-    let debugSql = '';
-    try {
-      debugSql = parameterizedQueryToSql({ sql: query, params: query_params });
-    } catch (e) {
-      debugSql = query;
-    }
-    let _url = this.host;
-
-    // eslint-disable-next-line no-console
-    console.log('--------------------------------------------------------');
-    // eslint-disable-next-line no-console
-    console.log('Sending Query:', debugSql);
-    // eslint-disable-next-line no-console
-    console.log('--------------------------------------------------------');
-
-    let clickhouse_settings = structuredClone(
-      external_clickhouse_settings || {},
-    );
-    if (clickhouse_settings?.max_rows_to_read && this.maxRowReadOnly) {
-      delete clickhouse_settings['max_rows_to_read'];
-    }
-
-    const createClient = await getJSNativeCreateClient();
-
-    if (isBrowser) {
-      clickhouse_settings = {
-        date_time_output_format: 'iso',
-        wait_end_of_query: 0,
-        cancel_http_readonly_queries_on_client_close: 1,
-        ...clickhouse_settings,
-      };
-      const http_headers = {
-        ...(connectionId && connectionId !== 'local'
-          ? { 'x-hyperdx-connection-id': connectionId }
-          : {}),
-      };
-      let myFetch: typeof fetch;
-      const isLocalMode = this.username != null && this.password != null;
-      if (isLocalMode) {
-        myFetch = localModeFetch;
-        clickhouse_settings.add_http_cors_header = 1;
-      } else {
-        _url = `${window.origin}${this.host}`; // this.host is just a pathname in this scenario
-        myFetch = standardModeFetch;
-      }
-
-      const url = new URL(_url);
-      const clickhouseClient = createClient({
-        url: url.origin,
-        pathname: url.pathname,
-        http_headers,
-        clickhouse_settings,
-        username: this.username ?? '',
-        password: this.password ?? '',
-        // Disable keep-alive to prevent multiple concurrent dashboard requests from exceeding the 64KB payload size limit.
-        keep_alive: {
-          enabled: false,
-        },
-        fetch: myFetch,
-      });
-      return clickhouseClient.query<Format>({
-        query,
-        query_params,
-        format,
-        abort_signal,
-        clickhouse_settings,
-        query_id: queryId,
-      }) as Promise<BaseResultSet<ReadableStream, Format>>;
-    } else if (isNode) {
-      const _client = createClient({
-        url: this.host,
-        username: this.username,
-        password: this.password,
-      });
-
-      // TODO: Custom error handling
-      return _client.query<Format>({
-        query,
-        query_params,
-        format,
-        abort_signal,
-        clickhouse_settings: {
-          date_time_output_format: 'iso',
-          wait_end_of_query: 0,
-          cancel_http_readonly_queries_on_client_close: 1,
-          ...clickhouse_settings,
-        },
-        query_id: queryId,
-      }) as unknown as Promise<BaseResultSet<ReadableStream, Format>>;
-    } else {
-      throw new Error(
-        'ClickhouseClient is only supported in the browser or node environment',
-      );
-    }
-  }
+  protected abstract __query<Format extends DataFormat>(
+    inputs: QueryInputs<Format>,
+  ): Promise<BaseResultSet<ReadableStream, Format>>;
 
   // TODO: only used when multi-series 'metrics' is selected (no effects on the events chart)
   // eventually we want to generate union CTEs on the db side instead of computing it on the client side
@@ -680,28 +542,6 @@ export class ClickhouseClient {
     throw new Error('No result sets');
   }
 }
-
-export const testLocalConnection = async ({
-  host,
-  username,
-  password,
-}: {
-  host: string;
-  username: string;
-  password: string;
-}): Promise<boolean> => {
-  try {
-    const client = new ClickhouseClient({ host, username, password });
-    const result = await client.query({
-      query: 'SELECT 1',
-      format: 'TabSeparatedRaw',
-    });
-    return result.text().then(text => text.trim() === '1');
-  } catch (e) {
-    console.warn('Failed to test local connection', e);
-    return false;
-  }
-};
 
 export const tableExpr = ({
   database,
